fixing update
