--- dotnet/CodeSnippets/MakeSuggestion.txt
@@ -1,54 +1,0 @@
-// This funtion goes in the Controllers/HomeController.cs file under the "helper functions" comment
-
-public Result MakeSuggestion(string emotion)
-{
-    var link = "";
-    var description = "";
-    // Decide which image to use based on the emotion in the picture
-    switch (emotion)
-    {
-        case "anger":
-            link = "/images/AngryDwight.jpg";
-            description = "You look just as angry as Dwight!";
-            break;
-        case "contempt":
-            link = "/images/Angela.jpg";
-            description = "Your contempt level is reaching Angela levels!";
-            break;
-        case "disgust":
-            link = "/images/Kelly.gif";
-            description = "You look just as disgusted as Kelly!";
-            break;
-        case "fear":
-            link = "/images/Michael.jpg";
-            description = "You might just be just as fearful and superstitous as Michael!";
-            break;
-        case "happiness":
-            link = "/images/JimAndPam.png";
-            description = "You look just as happy as Jim and Pam together!";
-            break;
-        case "neutral":
-            link = "/images/Stanley.jpg";
-            description = "Looking just as neutral as Stanley staring at the camera.";
-            break;
-        case "sadness":
-            link = "/images/SadDwight.jpg";
-            description = "Looking as sad as Dwight today!";
-            break;
-        case "surprise":
-            link = "/images/Jim.jpg";
-            description = "You look just as suprised as Jim walking in on Dwight's birthday suprise.";
-            break;
-        case "":
-            link = "/images/Error.jpg";
-            description = "Oops something went wrong! Please make sure that you submitted the correct image link and that your face is both promienent in the image and unobstructed. Submit another link to try again!";
-            break;
-    }
-    // Store suggestion
-    Result suggestion = new Result()
-    {
-        photo = link,
-        description = description
-    };
-    return suggestion;
-}
