--- dotnet/CodeSnippets/TopEmotion.txt
@@ -1,44 +1,0 @@
-// This function goes in the Models/EmotionSet.cs file inside of the Scores class
-
-public string Max()
-{
-    // Store all emotions for easier comparison
-    double[] scores = new double[] { anger,
-                                   contempt,
-                                   disgust,
-                                   fear,
-                                   happiness,
-                                   neutral,
-                                   sadness,
-                                   surprise };
-    // Find top emotion
-    int topEmotion = 0;
-    for (int i = 0; i < scores.Length; i++)
-    {
-        if (scores[i] > scores[topEmotion])
-        {
-            topEmotion = i;
-        }
-    }
-    // Return top emotion
-    switch (topEmotion)
-    {
-        case 0:
-            return "anger";
-        case 1:
-            return "contempt";
-        case 2:
-            return "disgust";
-        case 3:
-            return "fear";
-        case 4:
-            return "happiness";
-        case 5:
-            return "neutral";
-        case 6:
-            return "sadness";
-        case 7:
-            return "surprise";
-    }
-    return null;
-}
