--- dotnet/appsettings.json
@@ -1,13 +1,0 @@
-{
-  "ConnectionStrings": {
-    "DefaultConnection": "Data Source=WebApplication.db"
-  },
-  "Logging": {
-    "IncludeScopes": false,
-    "LogLevel": {
-      "Default": "Debug",
-      "System": "Information",
-      "Microsoft": "Information"
-    }
-  }
-}
