+++ dotnet/bin/Debug/netcoreapp1.1/dotnetporjects.runtimeconfig.dev.json
@@ -1,0 +1,7 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "C:\\Users\\t-clmcca\\.nuget\\packages"
+    ]
+  }
+}
