+++ dotnet/bin/Debug/netcoreapp1.1/dotnetporjects.runtimeconfig.json
@@ -1,0 +1,11 @@
+{
+  "runtimeOptions": {
+    "framework": {
+      "name": "Microsoft.NETCore.App",
+      "version": "1.1.2"
+    },
+    "configProperties": {
+      "System.GC.Server": true
+    }
+  }
+}
