--- dotnet/bower.json
@@ -1,10 +1,0 @@
-{
-  "name": "webapplication",
-  "private": true,
-  "dependencies": {
-    "bootstrap": "3.3.6",
-    "jquery": "2.2.3",
-    "jquery-validation": "1.15.0",
-    "jquery-validation-unobtrusive": "3.2.6"
-  }
-}
