--- dotnet/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "webapplication",
-  "version": "0.0.0",
-  "private": true,
-  "devDependencies": {
-    "gulp": "3.9.1",
-    "gulp-concat": "2.6.0",
-    "gulp-cssmin": "0.1.7",
-    "gulp-uglify": "1.5.3",
-    "rimraf": "2.5.2"
-  }
-}
